Added api hooks for beer recipes.

+++ api/ingredient/delete.php
@@ -3,6 +3,10 @@
 require '../init.php';
 require '../tools.php';
 
+if(!isLoggedIn()) {
+    fail("Only logged in users can delete ingredients");
+}
+
 $id = htmlspecialchars($_POST['id']);
 
 $query = 'DELETE FROM ingredient WHERE id = ? LIMIT 1';
@@ -13,6 +17,6 @@ $stmt->bind_param("d", $id);
 
 if($stmt->execute()) {
     success();
-} else {
-    fail($stmt->error);
 }
+
+fail($stmt->error);
